refactor(artwork): merge item artwork state into ArtworkRepository
